PDF AI/requirements.txt
Flask==2.0.3
PyMuPDF==1.18.19
openai==0.27.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2  
python-dotenv==0.19.2